extension for ESP boards

+++ SensorDataBuffer.h
@@ -85,10 +85,10 @@ public:
 		max_y=_data_y[_size-1];
 		for(int i=_startIndex;i<_size-1;i++)
 		{
-			min_x=min(min_x,_data_x[i]);
-			max_x=max(max_x,_data_x[i]);
-			min_y=min(min_y,_data_y[i]);
-			max_y=max(max_y,_data_y[i]);
+			min_x=min(min_x,( float)_data_x[i]);
+			max_x=max(max_x, (float)_data_x[i]);
+			min_y=min(min_y, (float)_data_y[i]);
+			max_y=max(max_y, (float)_data_y[i]);
 		}
 		min_x+=_offset_x;
 		max_x+=_offset_x;
